move "verifyValidation()" to model.
