+++ apis/userValidation/model/userValidation.model.ts
@@ -1,13 +1,22 @@
+import * as express from 'express';
 import { emailUtil } from '../../../packages/utils/email.util';
 import { encriptionPw } from '../../../packages/utils/encryption.util';
 import { mysqlUtil } from '../../../packages/utils/mysql.util';
+import { user } from '../../user/model/user.model';
 import smtpTransport = emailUtil.smtpTransport;
 
 const pool = mysqlUtil.pool;
 
 export class UserValidation {
+	public userValidationModel: express.Router = express.Router();
 	constructor() {
+		this.router();
 	}
+
+	public router() {
+        this.userValidationModel.get('/userValidation/verify/:uuid', verifyValidation);
+	}
+
 
 	/**
 	 * model: userValidation 생성
@@ -299,4 +308,57 @@ export class UserValidation {
 	}
 }
 
+
+
 export const userValidation: UserValidation = new UserValidation();
+
+
+
+/**
+ * route: 인증코드 검증
+ * @param req
+ * @param res
+ * @returns {Promise<void>}
+ */
+async function verifyValidation(req, res): Promise<void> {
+    try {
+        if (req.protocol == 'http') {
+
+            let verifiedUuid: any = req.params.uuid;
+
+            let uvUserId = await userValidation.getUserIdData(verifiedUuid);
+            uvUserId = JSON.stringify(uvUserId);
+
+            /** 해당 데이터가 없으면 [] */
+            if (uvUserId == '[]') {
+                res.send('Unvalidated code Error!!');
+            }
+
+            let userId = uvUserId.split('"')[3];
+
+            let uvUpdatedAt = await userValidation.getUpdatedAt(userId);
+            uvUpdatedAt = JSON.stringify(uvUpdatedAt);
+            uvUpdatedAt = uvUpdatedAt.split('"')[3];
+
+            let uvDate = uvUpdatedAt.split('T')[0].split('-');
+            let uvYearUpdatedAt = parseInt(uvDate[0]);
+            let uvMonthUpdatedAt = parseInt(uvDate[1]);
+            let uvDayUpdatedAt = parseInt(uvDate[2]);
+
+            if (user.isValidOnDate(uvYearUpdatedAt, uvMonthUpdatedAt, uvDayUpdatedAt)) {
+                await userValidation.updateIsValidation(userId);
+                await userValidation.deleteUsersValidation(userId);
+                await user.updateIsValidation(userId);
+                res.send('Email is been Successfully verified');
+            }
+            else {
+                res.send('validation date expired.')
+            }
+        }
+        else {
+            res.send('Request is from unknown source');
+        }
+    } catch (err) {
+        res.send(err);
+    }
+}
