+++ apis/userValidation/route/userValidation.route.ts
@@ -17,7 +17,7 @@ export class UserValidationRoutes {
 		this.userValidationRouter.post('/userValidation/userId/:userId/userPw', checkUserPw);
 		this.userValidationRouter.get('/userValidation/sendPasswordMail/:userId', sendPasswordMail);
 		this.userValidationRouter.post('/userValidation/sendValidationMail', sendValidationMail);
-		this.userValidationRouter.get('/userValidation/verify/:uuid', verifyValidation);
+		//this.userValidationRouter.get('/userValidation/verify/:uuid', verifyValidation);
 	}
 }
 
@@ -229,56 +229,6 @@ async function sendValidationMail(req, res): Promise<void> {
 					message: 'sendValidationMail: 50000'
 				});
 		}
-	}
-}
-
-/**
- * route: 인증코드 검증
- * @param req
- * @param res
- * @returns {Promise<void>}
- */
-async function verifyValidation(req, res): Promise<void> {
-	try {
-		if (req.protocol == 'http') {
-
-			let verifiedUuid: any = req.params.uuid;
-
-			let uvUserId = await userValidation.getUserIdData(verifiedUuid);
-			uvUserId = JSON.stringify(uvUserId);
-
-			/** 해당 데이터가 없으면 [] */
-			if (uvUserId == '[]')
-			{
-				res.send('Unvalidated code Error!!');
-			}
-
-			let userId = uvUserId.split('"')[3];
-
-			let uvUpdatedAt = await userValidation.getUpdatedAt(userId);
-			uvUpdatedAt = JSON.stringify(uvUpdatedAt);
-			uvUpdatedAt = uvUpdatedAt.split('"')[3];
-
-			let uvDate = uvUpdatedAt.split('T')[0].split('-');
-			let uvYearUpdatedAt = parseInt(uvDate[0]);
-			let uvMonthUpdatedAt = parseInt(uvDate[1]);
-			let uvDayUpdatedAt = parseInt(uvDate[2]);
-
-			if (user.isValidOnDate(uvYearUpdatedAt, uvMonthUpdatedAt, uvDayUpdatedAt)) {
-				await userValidation.updateIsValidation(userId);
-				await userValidation.deleteUsersValidation(userId);
-				await user.updateIsValidation(userId);
-				res.send('Email is been Successfully verified');
-			}
-			else {
-				res.send('validation date expired.')
-			}
-		}
-		else {
-			res.send('Request is from unknown source');
-		}
-	} catch (err) {
-		res.send(err);
 	}
 }
